fix: 댓글 3계층 코드 추가, 에러 작성, 오타 수정

+++ api/controllers/comments.controller.js
@@ -1,26 +1,49 @@
 const CommentsService = require("../services/comments.service.js");
+const { makeError } = require("../error");
 
 class CommentsController {
+  commentsService = new CommentsService();
+
   //댓글 작성
   createComment = async (req, res, next) => {
-    const { postId } = req.params;
+    const { userId } = res.locals.user;
+    const { goodsId } = req.params;
     const { comment } = req.body;
-
-    const result = await this.CommentsService.createComment({
-      userId,
-      postId,
-      comment,
-    });
-    return res.status(201).json({ message: "댓글이 생성되었습니다." });
+    try {
+      if (!(goodsId && comment)) {
+        throw new makeError({
+          message: "적절하지 않은 파라미터 요청입니다.",
+          code: 400,
+        });
+      }
+      const result = await this.CommentsService.createComment({
+        userId,
+        goodsId,
+        comment,
+      });
+      return res.status(201).json(result);
+    } catch (err) {
+      next(err);
+    }
   };
 
   //댓글 조회
   getComments = async (req, res, next) => {
-    const { postId } = req.parmas;
-    const result = await this.CommentsService.getComments({
-      postId,
-    });
-    return res.status(200).json({ data: result });
+    const { goodsId } = req.parmas;
+    try {
+      if (!goodsId) {
+        throw new makeError({
+          message: "적절하지 않은 파라미터 요청입니다.",
+          code: 400,
+        });
+      }
+      const result = await this.CommentsService.getComments({
+        goodsId,
+      });
+      return res.status(200).json(result);
+    } catch (err) {
+      next(err);
+    }
   };
 
   //댓글 수정
@@ -29,12 +52,28 @@ class CommentsController {
     const { commentId } = req.params;
     const { comment } = req.body;
 
-    const result = await this.CommentsService.editComments({
-      userId,
-      commentId,
-      comment,
-    });
-    return res.status(200).json({ message: "댓글이 수정되었습니다." });
+    try {
+      if (!(commentId && comment)) {
+        throw new makeError({
+          message: "적절하지 않은 파라미터 요청입니다.",
+          code: 400,
+        });
+      }
+      const result = await this.CommentsService.editComments({
+        userId,
+        commentId,
+        comment,
+      });
+      if (!result) {
+        throw new makeError({
+          messasge: "댓글조회에 실패했습니다.",
+          code: 404,
+        });
+      }
+      return res.status(200).json(result);
+    } catch (err) {
+      next(err);
+    }
   };
 
   //댓글 삭제
@@ -42,11 +81,21 @@ class CommentsController {
     const { userId } = res.locals.user;
     const { commentId } = res.params;
 
-    const result = await this.CommentsService.deleteComment({
-      userId,
-      commentId,
-    });
-    return res.status(200).json({ message: "댓글이 삭제되었습니다." });
+    try {
+      if (!coomentId) {
+        throw new makeError({
+          message: "적절하지 않은 파라미터 요청입니다.",
+          code: 400,
+        });
+      }
+      const result = await this.CommentsService.deleteComment({
+        userId,
+        commentId,
+      });
+      return res.status(200).json(result);
+    } catch (err) {
+      next(err);
+    }
   };
 }
